pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tafelmusik"
dynamic = ["version"]
description = "Collaborative editing layer for Claude and Sameer"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pycrdt>=0.12.50",
    "pycrdt-store>=0.1.0",
    "httpx-ws>=0.7.0",
    "starlette>=0.46.0",
    "uvicorn[standard]>=0.34.0",
    "mcp>=1.0.0",
]

[dependency-groups]
dev = [
    "hypothesis>=6.151.10",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9.0",
]

[tool.hatch.version]
path = ".claude-plugin/plugin.json"
pattern = '"version":\s*"(?P<version>[^"]+)"'

[tool.hatch.build.targets.wheel]
packages = ["src/tafelmusik"]

[tool.pytest.ini_options]
testpaths = ["src"]
asyncio_mode = "auto"
python_files = ["*_test.py"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
